+++ back/db.json
@@ -15,14 +15,14 @@
       "likes": "10"
     },
     {
-      "id": 3,
+      "id": 1,
       "title": "변수명의 차이",
       "content": "const, let은 변수 호이스팅이 일어나지 않는것처럼 보이지만",
       "category": "js",
       "likes": "17"
     },
     {
-      "id": 4,
+      "id": 2,
       "title": "배열의 반복",
       "content": "for뿐만아니라 forEach, map을 통하여 배열의 반복을",
       "category": "js",
